% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evds_functions.R
\name{all_series}
\alias{all_series}
\title{List of series}
\usage{
all_series()
}
\description{
Show All Series Function
}
\details{
This function allows you to preview all the series on the EVDS website.
It also creates a xlxs file where you can look at the data in another format.
This function helps user to determine the series they want to work on.
}
