% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evds_functions.R
\name{get_all_series}
\alias{get_all_series}
\title{Get All Series}
\usage{
get_all_series(export_to_xlsx = FALSE, export_path = "")
}
\arguments{
\item{export_to_xlsx}{Exports the output to xlsx.}

\item{export_path}{Path to export.}
}
\description{
This function allows you to preview all the series on the EVDS website. It also creates a xlxs file where you can look at the data in another format. This function helps user to determine the series they want to work on.
}
