% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evds_functions.R
\name{get_data_from_phrase}
\alias{get_data_from_phrase}
\title{Common Wrapper for Data Parsing}
\usage{
get_data_from_phrase(the_phrase)
}
\arguments{
\item{the_phrase}{JSON string to send via GET parameter.}
}
