% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evds_functions.R
\name{get_series}
\alias{get_series}
\title{Get Series}
\usage{
get_series(series = c("TP.DK.USD.A", "TP.DK.EUR.A"),
  start_date = "01-10-2017", end_date = "01-11-2017")
}
\arguments{
\item{series, start_date, end_date}{The desired series should be put in the series in a vector form. start_date and end_date are in form dd-mm-YYYY.}
}
\description{
Get series Function
}
\details{
This function allows you to get series you want to work on from the EVDS website. You can also choose the dates you want to get the data from.
}
\examples{
get_series(series=c("",""), start_date = "dd-mm-YYYY", end_date = "dd-mm-YYYY")
}
