% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evds_functions.R
\name{set_evds_key}
\alias{set_evds_key}
\title{Set API KEY}
\usage{
set_evds_key(api_key)
}
\arguments{
\item{api_key}{Your API key from the EVDS website.}
}
\description{
This function allows you to set your API key. You will need this key to export data from the EVDS website.
}
\examples{
set_evds_key("YOURAPIKEY")
}
