% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evds_functions.R
\name{show_serie_lists}
\alias{show_serie_lists}
\title{List of series based on a category code or name}
\usage{
show_serie_lists(ccode)
}
\arguments{
\item{ccode}{Category code}
}
\description{
Show Data Groups Function
}
\details{
This function allows you to preview the series you want to work on from the EVDS website.
}
