% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evds_functions.R
\name{show_categories}
\alias{show_categories}
\title{Data Categories}
\usage{
show_categories()
}
\description{
Show Data Categories Function
}
\details{
This function allows you to preview the category of series you want to work on from the EVDS website.
}
