% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evds_functions.R
\name{show_serie_lists}
\alias{show_serie_lists}
\title{List of series based on a category code or name}
\usage{
show_serie_lists(ccode)
}
\arguments{
\item{ccode}{Category code}
}
\description{
This function allows you to preview the series from a specific data group you want to work on from the EVDS website. You can get the name of the datagroups by using `show_categories` function.
}
\examples{
\dontrun{
show_serie_lists("bie_fafaiz")
}
}
